% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_kza_filter.R
\name{apply_kza_filter}
\alias{apply_kza_filter}
\title{Function to apply the adaptive Kolmogorov-Zurbenko (KZA) filter to a time 
series, usually used for discontinuity identification.}
\usage{
apply_kza_filter(df, window = 365, k = 3, impute_tails = FALSE)
}
\arguments{
\item{df}{Data frame containing daily time series observations.}

\item{window}{Window for K-Z filters.}

\item{k}{Number of iterations.}

\item{impute_tails}{Should the leading and trailing values in the time series
be imputed after the K-Z filters have been applied?}
}
\value{
Tibble.
}
\description{
Function to apply the adaptive Kolmogorov-Zurbenko (KZA) filter to a time 
series, usually used for discontinuity identification.
}
\author{
Sturat K. Grange
}
