% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_to_time_series.R
\name{data_frame_to_time_series}
\alias{data_frame_to_time_series}
\title{Function to convert a data frame to a time series (\code{ts}) object.}
\usage{
data_frame_to_time_series(df, variable = "value", period = "month")
}
\arguments{
\item{df}{Data frame containing a date variable.}

\item{variable}{Variable in \code{df} to create a time series object with.}

\item{period}{Period/interval to make time series. Currently, only 
\code{"month"} and \code{"day"} are supported.}
}
\value{
A \code{ts} object.
}
\description{
Function to convert a data frame to a time series (\code{ts}) object.
}
\author{
Stuart K. Grange
}
