% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose_with_kz_filters.R
\name{decompose_with_kz_filters}
\alias{decompose_with_kz_filters}
\title{Function to decompose a time series with Kolmogorov-Zurbenko (K-Z) filters.}
\usage{
decompose_with_kz_filters(df)
}
\arguments{
\item{df}{A tibble containing time series observations.}
}
\value{
Tibble.
}
\description{
Function to decompose a time series with Kolmogorov-Zurbenko (K-Z) filters.
}
\author{
Stuart K. Grange
}
