% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_anomalies_with_shesd.R
\name{detect_anomalies_with_shesd}
\alias{detect_anomalies_with_shesd}
\title{Function to detect anomalies with the seasonal hybrid extreme studentised
deviate (SHESD) model.}
\usage{
detect_anomalies_with_shesd(
  df,
  max_fraction = 0.02,
  direction = "both",
  long_term = FALSE,
  only_last = NULL,
  invalidate = FALSE
)
}
\arguments{
\item{df}{Input time series stored in a data frame or tibble with \code{date}
and \code{value} variables.}

\item{max_fraction}{Maximum fraction (from all observations in \code{df}) 
anomalies that SHESD will detect.}

\item{direction}{Directionality of the anomalies to be detected. One of 
\code{pos}, \code{neg}, or \code{both}.}

\item{long_term}{Should a more efficient algorithm which employs piecewise 
approximation be applied to the time series?}

\item{only_last}{Detect anomalies only within the last day or hour in the 
time series. One of \code{NULL}, \code{day}, or \code{hr}.}

\item{invalidate}{Should observations in \code{df} be invalidated if detected
as anomalies?}
}
\value{
Tibble.
}
\description{
Function to detect anomalies with the seasonal hybrid extreme studentised
deviate (SHESD) model.
}
\seealso{
\url{https://github.com/hrbrmstr/AnomalyDetection}
}
\author{
Stuart K. Grange
}
