% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_time_series.R
\name{forecast_time_series}
\alias{forecast_time_series}
\title{Function to forecast observations in a time series.}
\usage{
forecast_time_series(df, method = c("arima", "exponential"), n = 10, ci = 80)
}
\arguments{
\item{df}{Input tibble containing time series data. \code{df} requires a 
\code{date} and \code{value} variables.}

\item{method}{Modelling method to use. Either \code{exponential} or 
\code{arima} for exponential smoothing or autoregressive integrated moving
average (ARIMA).}

\item{n}{Number of steps to forecast to.}

\item{ci}{Confidence level for prediction intervals.}
}
\value{
Tibble.
}
\description{
Function to forecast observations in a time series.
}
\author{
Stuart K. Grange
}
