% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saq_trend_plot.R
\name{saq_trend_plot}
\alias{saq_trend_plot}
\title{Function to build a trend plot from the output of \code{\link{saq_trend_test}}.}
\usage{
saq_trend_plot(
  df,
  df_tests,
  label = TRUE,
  round = 3,
  y_location = 1,
  include_n = TRUE,
  facet_variable = NA,
  scales = "fixed",
  ylim = c(NA, NA),
  colour = "#FCA50A",
  x_label_rotate = 0,
  parse_facet_label = FALSE,
  point_size = 1.5
)
}
\arguments{
\item{df}{Input data frame, generally containing deseasonalised observations
from \code{\link{saq_trend_test}}.}

\item{df_tests}{Input data frame, generally containing trend tests calculated
by \code{\link{saq_trend_test}}.}

\item{label}{Should the plots have a trend magnitude label added?}

\item{round}{If \code{label} is \code{TRUE}, how many digits should the trend
estimated be rounded too?}

\item{y_location}{If \code{label} is \code{TRUE}, where should the label be
positioned? In a decimal of the maximum y value.}

\item{include_n}{If \code{label} is \code{TRUE}, should the number of 
observations be included in the label?}

\item{facet_variable}{What variable in \code{df} and \code{df_tests} should
be used for faceting?}

\item{scales}{If \code{facet_variable} is \code{TRUE}, should the axes be 
free or fixed?}

\item{ylim}{A vector with the length of two for the y-axis limits. By default,
the y-axis will start at zero.}

\item{colour}{Either a colour of the plot's point and lines, or a variable in
\code{df} which the points' and lines' colours will be mapped to.}

\item{x_label_rotate}{Angle of the x-tick (date) labels. If overlapping 
labels are encountered, setting \code{x_angle} to \code{45} may help.}

\item{parse_facet_label}{Should the facet names be parsed as an expression?}

\item{point_size}{What size should the points be on the trend plot?}
}
\value{
\strong{ggplot2} plot.
}
\description{
Function to build a trend plot from the output of \code{\link{saq_trend_test}}.
}
\author{
Stuart K. Grange
}
