% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stl_to_data_frame.R
\name{stl_to_data_frame}
\alias{stl_to_data_frame}
\title{Function to convert a \code{stlplus} object to a data frame/tibble.}
\usage{
stl_to_data_frame(stl, date_round = TRUE, na_preserve = FALSE)
}
\arguments{
\item{stl}{An \code{stl} object.}

\item{date_round}{Should dates be rounded?}

\item{na_preserve}{Should observations which were missing with the values
also be propagated into the the deseasonalised components?}
}
\value{
Tibble.
}
\description{
Function to convert a \code{stlplus} object to a data frame/tibble.
}
\author{
Stuart K. Grange
}
