% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saq_trend_plot.R
\name{saq_trend_plot}
\alias{saq_trend_plot}
\title{Function to build a trend plot from the output of \code{\link{saq_trend_test}}.}
\usage{
saq_trend_plot(
  df,
  df_tests,
  label = TRUE,
  round = 3,
  y_location = 1,
  facet_variable = NA,
  scales = "fixed",
  colour = "#FCA50A",
  x_angle = NA
)
}
\arguments{
\item{df}{Input data frame, generally containing deseasonalised observations
from \code{\link{saq_trend_test}}.}

\item{df_tests}{Input data frame, generally containing trend tests calculated
by \code{\link{saq_trend_test}}.}

\item{label}{Should the plots have a trend magnitude label added?}

\item{round}{If \code{label} is \code{TRUE}, how many digits should the trend
estimated be rounded too?}

\item{y_location}{If \code{label} is \code{TRUE}, where should the label be
positioned? In a decimal of the maximum y value.}

\item{facet_variable}{What variable in \code{df} and \code{df_tests} should
be used for faceting?}

\item{scales}{If \code{facet_variable} is \code{TRUE}, should the axes be 
free or fixed?}

\item{colour}{Colour of the plot's point and lines.}

\item{x_angle}{Angle of the x-tick (date) labels. If overlapping labels are
encountered, setting \code{x_angle} to \code{45} may help.}
}
\value{
\strong{ggplot2} plot.
}
\description{
Function to build a trend plot from the output of \code{\link{saq_trend_test}}.
}
\author{
Stuart K. Grange
}
