% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saq_trend_test.R
\name{saq_trend_test}
\alias{saq_trend_test}
\title{Function to conduct a robust trend test.}
\usage{
saq_trend_test(df, by = as.character(), decompose = TRUE,
  window = 35, na_preserve = TRUE, alpha = 0.05,
  auto_correlation = FALSE, progress = FALSE)
}
\arguments{
\item{df}{Input data frame containing a time series.}

\item{by}{Which variables should be used as grouping variables?}

\item{decompose}{Should the time series be decomposed before the trend test? 
If this is \code{FALSE}, names in your inputs will be changed.}

\item{window}{Span (in lags) of the loess window for seasonal extraction. 
This should be an odd number.}

\item{na_preserve}{Should observations which were missing with the values 
also be propagated into the the deseasonalised components?}

\item{alpha}{Confidence interval of the slope, default is 0.05 for 95 % 
confidence intervals.}

\item{auto_correlation}{Should auto correlation be considered in the 
estimates?}

\item{progress}{Should a progress bar be displayed for the trend test 
calculations?}
}
\value{
Named list.
}
\description{
\code{saq_trend_test} can conduct the trend tests in parallel if a 
\strong{futures} backend has been registered.
}
\author{
Stuart K. Grange
}
