% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saq_trend_test.R
\name{saq_trend_test}
\alias{saq_trend_test}
\alias{saq_trend_test_nested}
\title{Function to conduct a robust trend test using the Theil-Sen estimator.}
\usage{
saq_trend_test(
  df,
  by = as.character(),
  decompose = TRUE,
  window = 35,
  na_preserve = TRUE,
  alpha = 0.05,
  auto_correlation = FALSE,
  period = "month",
  progress = FALSE
)

saq_trend_test_nested(
  df_nest,
  decompose = FALSE,
  window = 35,
  na_preserve = TRUE,
  alpha = 0.05,
  auto_correlation = FALSE,
  period = "month",
  progress = FALSE
)
}
\arguments{
\item{df}{Input data frame containing a time series.}

\item{by}{Which variables should be used as grouping variables?}

\item{decompose}{Should the time series be decomposed before the trend test? 
If this is \code{FALSE}, names in your input will be changed.}

\item{window}{Span (in lags) of the loess window for seasonal extraction. 
This should be an odd number.}

\item{na_preserve}{Should observations which were missing with the values 
also be propagated into the deseasonalised components?}

\item{alpha}{Confidence interval of the slope, default is 0.05 for 95 % 
confidence intervals.}

\item{auto_correlation}{Should auto correlation be considered in the 
estimates?}

\item{period}{Period of input time series. Default is \code{"month"} but can
also be \code{"year"}.}

\item{progress}{Should a progress bar be displayed for the trend test 
calculations?}

\item{df_nest}{Nested tibble including a nested observational variable named
\code{observations}.}
}
\value{
Named list or nested tibble.
}
\description{
\code{saq_trend_test} can conduct the trend tests in parallel if a 
\strong{futures} backend has been registered.
}
\author{
Stuart K. Grange
}
